refactor(*): code clean for quota ctrl knob
