--- pkg/agent/sysadvisor/plugin/qosaware/server/cpu_server_unsupported.go
@@ -1,29 +1,0 @@
-//go:build !linux
-// +build !linux
-
-/*
-Copyright 2022 The Katalyst Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package server
-
-import (
-	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/advisorsvc"
-	"github.com/kubewharf/katalyst-core/pkg/agent/sysadvisor/types"
-)
-
-func (cs *cpuServer) assembleCgroupConfig(advisorResp *types.InternalCPUCalculationResult) (extraEntries []*advisorsvc.CalculationInfo) {
-	return []*advisorsvc.CalculationInfo{}
-}
